refactor: project 파일명 변경
