+++ src/main/java/com/example/cacheproject/collection/repository/CsvDataRepository.java
@@ -1,0 +1,7 @@
+package com.example.cacheproject.collection.repository;
+
+import com.example.cacheproject.collection.entity.CsvData;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CsvDataRepository extends JpaRepository<CsvData, Long> {
+}
