+++ src/main/java/com/example/cacheproject/config/RestTemplateConfig.java
@@ -1,0 +1,13 @@
+package com.example.cacheproject.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RestTemplateConfig {
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
