+++ src/main/java/com/example/cacheproject/config/XmlMapperConfig.java
@@ -1,0 +1,14 @@
+package com.example.cacheproject.config;
+
+import com.fasterxml.jackson.dataformat.xml.XmlMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class XmlMapperConfig {
+
+    @Bean
+    public XmlMapper xmlMapper() {
+        return new XmlMapper();
+    }
+}
