+++ src/main/java/com/example/cacheproject/exception/BadRequestException.java
@@ -1,0 +1,12 @@
+package com.example.cacheproject.exception;
+
+public class BadRequestException extends CustomException {
+
+    public BadRequestException() {
+        super(ErrorCode.BAD_REQUEST);
+    }
+
+    public BadRequestException(String message) {
+        super(ErrorCode.BAD_REQUEST, message);
+    }
+}
