+++ src/main/java/com/example/cacheproject/exception/DataIntegrityException.java
@@ -1,0 +1,10 @@
+package com.example.cacheproject.exception;
+
+public class DataIntegrityException extends CustomException {
+
+    public DataIntegrityException() { super(ErrorCode.DATA_INCONSISTENCY); }
+
+    public DataIntegrityException(String message) {
+        super(ErrorCode.DATA_INCONSISTENCY, message);
+    }
+}
