+++ src/main/java/com/example/cacheproject/exception/ForbiddenException.java
@@ -1,0 +1,12 @@
+package com.example.cacheproject.exception;
+
+public class ForbiddenException extends CustomException {
+
+    public ForbiddenException() {
+        super(ErrorCode.FORBIDDEN);
+    }
+
+    public ForbiddenException(String message) {
+        super(ErrorCode.FORBIDDEN,message);
+    }
+}
