+++ src/main/java/com/example/cacheproject/exception/NotFoundException.java
@@ -1,0 +1,13 @@
+package com.example.cacheproject.exception;
+
+public class NotFoundException extends CustomException {
+
+    public NotFoundException() {
+        super(ErrorCode.NOT_FOUND);
+    }
+
+    public NotFoundException(String message) {
+        super(ErrorCode.NOT_FOUND, message);
+    }
+}
+
