+++ src/main/java/com/example/cacheproject/exception/UnauthorizedException.java
@@ -1,0 +1,12 @@
+package com.example.cacheproject.exception;
+
+public class UnauthorizedException extends CustomException {
+
+    public UnauthorizedException() {
+        super(ErrorCode.AUTHORIZATION);
+    }
+
+    public UnauthorizedException(String message) {
+        super(ErrorCode.AUTHORIZATION, message);
+    }
+}
