+++ src/main/java/com/example/cacheproject/openapi/repository/OpenApiRepository.java
@@ -1,0 +1,7 @@
+package com.example.cacheproject.openapi.repository;
+
+import com.example.cacheproject.openapi.entity.OpenApi;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OpenApiRepository extends JpaRepository<OpenApi, Long> {
+}
