+++ src/main/java/com/example/cacheproject/response/MessageResponse.java
@@ -1,0 +1,19 @@
+package com.example.cacheproject.response;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.Getter;
+
+@Getter
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class MessageResponse {
+    String message;
+
+    private MessageResponse(String message) {
+        this.message = message;
+    }
+
+    public static MessageResponse of(String message) {
+        return new MessageResponse(message);
+    }
+}
+
