--- src/main/java/com/example/plusproject/collection/fetchstatus/repository/CsvDataFetchStatusRepository.java
@@ -1,9 +1,0 @@
-package com.example.plusproject.collection.fetchstatus.repository;
-
-import com.example.plusproject.collection.fetchstatus.entity.CsvDataFetchStatus;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CsvDataFetchStatusRepository extends JpaRepository<CsvDataFetchStatus, Long> {
-
-    CsvDataFetchStatus findTopByOrderByUpdatedAtDesc();  // 가장 최근에 업데이트된 레코드 찾기
-}
