--- src/main/java/com/example/plusproject/config/RestTemplateConfig.java
@@ -1,13 +1,0 @@
-package com.example.plusproject.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-@Configuration
-public class RestTemplateConfig {
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
