--- src/main/java/com/example/plusproject/config/XmlMapperConfig.java
@@ -1,14 +1,0 @@
-package com.example.plusproject.config;
-
-import com.fasterxml.jackson.dataformat.xml.XmlMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class XmlMapperConfig {
-
-    @Bean
-    public XmlMapper xmlMapper() {
-        return new XmlMapper();
-    }
-}
