--- src/main/java/com/example/plusproject/exception/BadRequestException.java
@@ -1,12 +1,0 @@
-package com.example.plusproject.exception;
-
-public class BadRequestException extends CustomException {
-
-    public BadRequestException() {
-        super(ErrorCode.BAD_REQUEST);
-    }
-
-    public BadRequestException(String message) {
-        super(ErrorCode.BAD_REQUEST, message);
-    }
-}
