--- src/main/java/com/example/plusproject/exception/DataIntegrityException.java
@@ -1,10 +1,0 @@
-package com.example.plusproject.exception;
-
-public class DataIntegrityException extends CustomException {
-
-    public DataIntegrityException() { super(ErrorCode.DATA_INCONSISTENCY); }
-
-    public DataIntegrityException(String message) {
-        super(ErrorCode.DATA_INCONSISTENCY, message);
-    }
-}
