--- src/main/java/com/example/plusproject/exception/ForbiddenException.java
@@ -1,12 +1,0 @@
-package com.example.plusproject.exception;
-
-public class ForbiddenException extends CustomException {
-
-    public ForbiddenException() {
-        super(ErrorCode.FORBIDDEN);
-    }
-
-    public ForbiddenException(String message) {
-        super(ErrorCode.FORBIDDEN,message);
-    }
-}
