--- src/main/java/com/example/plusproject/exception/NotFoundException.java
@@ -1,13 +1,0 @@
-package com.example.plusproject.exception;
-
-public class NotFoundException extends CustomException {
-
-    public NotFoundException() {
-        super(ErrorCode.NOT_FOUND);
-    }
-
-    public NotFoundException(String message) {
-        super(ErrorCode.NOT_FOUND, message);
-    }
-}
-
