--- src/main/java/com/example/plusproject/openapi/fetchstatus/repository/OpenApiFetchStatusRepository.java
@@ -1,9 +1,0 @@
-package com.example.plusproject.openapi.fetchstatus.repository;
-
-
-import com.example.plusproject.openapi.fetchstatus.entity.OpenApiFetchStatus;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OpenApiFetchStatusRepository extends JpaRepository<OpenApiFetchStatus, Long> {
-    OpenApiFetchStatus findTopByOrderByUpdatedAtDesc();  // 가장 최근에 업데이트된 데이터를 찾는 메서드
-}
